Update start autopilot operators
